Add: separated graph packages
