+++ src/main/kotlin/model/abstractGraph/AbstractEdge.kt
@@ -1,4 +1,6 @@
-package model
+package model.abstractGraph
+
+import model.abstractGraph.AbstractVertex
 
 interface AbstractEdge {
     val id: Long
