+++ src/main/kotlin/model/abstractGraph/AbstractGraph.kt
@@ -1,4 +1,6 @@
-package model
+package model.abstractGraph
+
+import model.abstractGraph.AbstractVertex
 
 interface AbstractGraph {
     val vertices: Collection<AbstractVertex>
