+++ src/main/kotlin/view/graph/VertexView.kt
@@ -1,4 +1,4 @@
-package view
+package view.graph
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
@@ -13,7 +13,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.dp
-import viewModel.VertexViewModel
+import viewModel.graph.VertexViewModel
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
